edit-gender/requirements.txt
numpy==1.19.2
opencv-python==4.4.0.44
pandas==1.1.3
Pillow==8.0.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
